long no 2

+++ src/android/CodePush.java
@@ -442,7 +442,7 @@ public class CodePush extends CordovaPlugin {
             /* The application was resumed from the background. */
             /* Handle ON_NEXT_RESUME pending installations. */
             InstallOptions pendingInstall = this.codePushPackageManager.getPendingInstall();
-            int durationInBackground = (new Date().getTime() - lastPausedTimeMs) / 1000;
+            long durationInBackground = (new Date().getTime() - lastPausedTimeMs) / 1000;
             if (pendingInstall != null && InstallMode.ON_NEXT_RESUME.equals(pendingInstall.installMode) && durationInBackground >= pendingInstall.minimumBackgroundDuration) {
                 handleAppStart();
                 this.markUpdate();
